+++ .prettierignore
@@ -22,6 +22,7 @@ common/scripts
 coverage
 etc
 change
+temp
 
 # package specific files
 packages/web-components/**/*.spec.ts
