+++ package.json
@@ -91,10 +91,12 @@
     "@storybook/react": "^5.3.8",
     "@types/jest": "24.9.0",
     "@types/jest-axe": "^3.2.1",
+    "@types/node": "^10.3.2",
     "babel-jest": "24.9.0",
     "beachball": "^1.39.1",
     "cross-env": "^5.1.4",
     "danger": "^6.0.5",
+    "eslint": "^7.1.0",
     "gulp": "^4.0.2",
     "jest": "24.9.0",
     "jest-cli": "24.9.0",
@@ -103,6 +105,7 @@
     "lage": "0.27.0",
     "lerna": "^3.15.0",
     "lint-staged": "^10.2.9",
+    "prettier": "~1.19.1",
     "sass-loader": "^6.0.6",
     "satisfied": "^1.1.1",
     "syncpack": "^5.6.10",
