+++ packages/react-button/.eslintrc.json
@@ -1,4 +1,12 @@
 {
   "extends": ["plugin:@fluentui/eslint-plugin/react"],
-  "root": true
+  "root": true,
+  "overrides": [
+    {
+      "files": ["*.stories.{ts,tsx}"],
+      "rules": {
+        "no-alert": "warn"
+      }
+    }
+  ]
 }
