+++ packages/react-button/src/version.ts
@@ -2,4 +2,3 @@
 // The checked in version is a placeholder only and will not be updated.
 import { setVersion } from '@fluentui/set-version';
 setVersion('@fluentui/react-button', '0.0.0');
-
