+++ packages/react-internal/src/index.ts
@@ -71,4 +71,6 @@ export * from './Tooltip';
 export * from './Utilities';
 export * from './WindowProvider';
 
+// export * from './utilities/decorators/withViewport';
+
 import './version';
