+++ scripts/package.json
@@ -31,7 +31,6 @@
     "@types/jju": "^1.4.1",
     "@types/lerna-alias": "^3.0.0",
     "@types/lodash": "^4.14.149",
-    "@types/node": "^10.3.2",
     "@types/node-fetch": "^2.5.7",
     "@types/puppeteer": "1.12.3",
     "@types/read-pkg-up": "^6.0.0",
@@ -60,7 +59,6 @@
     "electron": "^8.4.1",
     "enzyme": "~3.10.0",
     "enzyme-adapter-react-16": "^1.15.0",
-    "eslint": "^7.1.0",
     "express": "^4.15.4",
     "extract-comments": "^1.0.0",
     "find-free-port": "~2.0.0",
@@ -93,7 +91,6 @@
     "postcss": "^7.0.30",
     "postcss-loader": "^3.0.0",
     "postcss-modules": "^2.0.0",
-    "prettier": "~1.19.1",
     "progress": "^2.0.3",
     "prompts": "^2.3.0",
     "puppeteer": "^1.13.0",
